+++ resources/views/partials/navigation.blade.php
@@ -12,7 +12,7 @@
             Dashboard
         </a>
 
-        <a href="{{ route('welcome') }}"
+        <a href="{{ route('catalogue') }}"
             class="group df_nav_links {{ Request::is('welcome') ? 'df_nav_links_active' : '' }}">
             @include('icons.catalogue', ['style'=>'h-5 w-5 mr-4'])
             Catalogue
@@ -25,8 +25,6 @@
             Courses
         </a>
         @endcan
-
-
 
         @can('crud_locations')
         <a href="{{ route('locations.index') }}" class="group df_nav_links">
@@ -74,11 +72,12 @@
         @endcan
 
 
-        <a href="{{ route('roles.index') }}"
-            class="group df_nav_links {{ Request::is('roles') ? 'df_nav_links_active' : '' }}">
+        <a href="{{ route('reports') }}"
+            class="group df_nav_links {{ Request::is('reports') ? 'df_nav_links_active' : '' }}">
             @include('icons.pie-chart', ['style'=>'h-5 w-5 mr-4'])
             Reports
         </a>
+
         <a href="{{ route('roles.index') }}"
             class="group df_nav_links {{ Request::is('roles') ? 'df_nav_links_active' : '' }}">
             @include('icons.settings-tool', ['style'=>'h-5 w-5 mr-4'])
@@ -89,10 +88,11 @@
     </div>
     <div class="mt-8">
         <!-- Secondary navigation -->
-        <h3 class="px-3 text-xs leading-4 font-semibold text-gray-500 uppercase tracking-wider" id="teams-headline">
+        {{-- <h3 class="px-3 text-xs leading-4 font-semibold text-gray-500 uppercase tracking-wider" id="teams-headline">
             Teams
-        </h3>
-        <div class="mt-1 space-y-1" role="group" aria-labelledby="teams-headline">
+        </h3> --}}
+
+        {{-- <div class="mt-1 space-y-1" role="group" aria-labelledby="teams-headline">
             <a href="#"
                 class="group flex items-center px-3 py-2 text-sm leading-5 font-medium text-gray-200 rounded-md hover:text-gray-900 hover:bg-gray-50 focus:outline-none focus:bg-gray-50 transition ease-in-out duration-150">
                 <span class="w-2.5 h-2.5 mr-4 bg-indigo-500 rounded-full"></span>
@@ -116,6 +116,6 @@
                     Urban
                 </span>
             </a>
-        </div>
+        </div> --}}
     </div>
 </nav>