+++ resources/views/partials/payment-method/bank-transfer.blade.php
@@ -1,0 +1,41 @@
+<div class="shadow overflow-hidden border-b border-gray-200 sm:rounded-lg px-5 pt-3 pb-8">
+    <table class="w-full">
+        <tr>
+            <td valign="top" class="font-semibold pl-2">
+                Bank
+            </td>
+            <td class="text-gray-700">
+                Migros Bank AG <br>
+                8001 Zürich
+            </td>
+        </tr>
+        <tr class="bg-gray-50">
+            <td class="font-semibold pl-2">
+                IBAN
+            </td>
+            <td class="text-gray-700">
+                CH65 0840 1000 0583 7151 6
+            </td>
+        </tr>
+        <tr>
+            <td class="font-semibold pl-2">Name</td>
+            <td class="text-gray-700">DANCEFLOOR Kouadio & Te</td>
+        </tr>
+        <tr valign="top" class="bg-gray-50">
+            <td class="font-semibold pl-2">Address</td>
+            <td class="text-gray-700">
+                c/o Diaz Eliseo <br>
+                Rue François-Durafour 14 <br>
+                1220 Les avanchets
+            </td>
+        </tr>
+        <tr>
+            <td class="font-bold px-2">Account Number</td>
+            <td class="text-gray-700">80-533-6</td>
+        </tr>
+        <tr class="bg-gray-50">
+            <td class="font-semibold pl-2">SWIFT / BIC</td>
+            <td class="text-gray-700">MIGR CH ZZ 12A</td>
+        </tr>
+    </table>
+</div>