+++ resources/views/partials/payment-method/cash.blade.php
@@ -1,0 +1,4 @@
+<div class="border rounded-lg p-5">
+    <h3 class="text-2xl">Pay by cash</h3>
+    <p class="mb-5">Please bring the exact amount, CHF {{ $total }}, to your next class and give it to your teacher</p>
+</div>