+++ resources/views/partials/payment-method/paypal.blade.php
@@ -1,0 +1,8 @@
+<div class="shadow overflow-hidden border-b border-gray-200 sm:rounded-lg px-5 pt-3 pb-8">
+    <h3 class="text-2xl">Paypal payment</h3>
+    <p class="mb-5">If you have a PayPal account you can make a transfer of CHF {{ $total }} to the following email</p>
+    <a href="mailto:dancefloor.geneva@gmail.com"
+        class="bg-red-700 px-4 py-2 text-white hover:bg-red-800 rounded-full my-5">
+        dancefloor.geneva@gmail.com
+    </a>
+</div>