+++ resources/views/partials/payment-method/revolut.blade.php
@@ -1,0 +1,7 @@
+<div class="shadow overflow-hidden border-b border-gray-200 sm:rounded-lg px-5 pt-3 pb-8">
+    <h3 class="text-2xl">Revolut money transfer</h3>
+    <p class="mb-5">If you have a revolut account, you can easily transfer CHF {{ $total }} to this phone number</p>
+    <a href="tel:+41 78 657 50 56" class="df-btn-primary">
+        +41 78 657 50 56
+    </a>
+</div>